feat: new method for adding booking slots
The endpoint in the Admin controller has been modified to support
adding new booking slots to the database. The admin now needs to pass
a parameter named "days" (a value between 1 and 30), which determines
the number of days for which the empty booking slots will be added.

By default, the new booking slots are added only between Monday and
Friday from 9 AM to 6 PM, with slots available every 30 minutes.

To implement this modification, the following changes were made:

  - Completely rebuild the "addEmptyBookingSlotsFor" method in Admin
    Controller

  - Added a new query named "findLatestDateTime()" in the Booking
    Repository, which retrieves the latest booking slot datetime from
    the repository.

  - Implemented a new method in the Booking Service named
    "latestDateTime", which returns the LocalDateTime of the last slot
     in the repository or the current LocalDateTime if the repository
     is empty.