fix: access to H2 console
The MvcRequestMatcher is designed to work with the DispatcherServlet's
rules for handling web requests. The H2 Console, however, doesn't
follow these rules because it's not managed by the DispatcherServlet.
However, special authentication rules are added using Ant-style
patterns to allow access to the H2 Console only for users with the
"ADMIN" role. Meanwhile, custom controllers in the application follow
the DispatcherServlet's rules for handling requests.