feat: new method for registering admin
The application now includes a new method for registering the admin.
Instead of registering a new admin, the application will add an initial
default admin to the database. The admin can then change the email and
password to their own.

This change enhances the application's security, as the previous
endpoint for registering users was publicly accessible.

Key points:
- Add DataLoaderAdmin to check for existing admin users and create
  a default admin if none are found.
- Add a new endpoint in AdminController (accessible only by admin)
  to update the admin's email and password.
- Remove the register admin endpoint.
- Update WebSecurityConfig.
- Add tests for the new endpoint.
- Update documentation.